IGNITE-25012 Extract GridToStringBuilder Unsafe code to interface
